mk/03_etl/requirements.txt
sqlalchemy
mongo
pandas
vaderSentiment
